Refactor MPI/OMP handling and streamline integration test infrastructure
- Introduce --omp-granularity to control rank/thread balance
  (e.g., N ranks with M threads vs N/2 ranks with 2*M threads)
- Add omp pragma for energy evaluation in Equations.cxx
- Simplify MPI communicator logic in atrip to support non-default rank mappings more cleanly
  (e.g. introduce a new communicator with node-aware rank distribution)
- Use consistent integer types (int) for all MPI-related values (rank, np, etc.) within atrip
- Replace size_t with int in tuples.hpp and slice.hpp to match MPI integer usage
- Overhaul integration-tests:
  - Centralize execution logic in launcher.sh
  - Use *.inp files for per-case configuration

+++ src/atrip/Slice.hpp
@@ -33,8 +33,8 @@ public:
 
   // [[file:~/cuda/atrip/atrip.org::*Location][Location:1]]
   struct Location {
-    size_t rank;
-    size_t source;
+    int rank;
+    int source;
   };
   // Location:1 ends here
